+++ Topicos de Programacion/Bibliotecas/Matrix.h
@@ -2,9 +2,9 @@
 #define MATRIX_H_INCLUDED
 #include <stdio.h>
 #include <stdlib.h>
-void   printMatrix (long long** matrix,  int rows,  int cols);
+void   printMatrix (int** matrix,  int rows,  int cols);
 void** createMatrix(int rows,  int cols,size_t elemSize);
-void   fillMatrixConditional(long long**matrix,  int rows,  int cols);
+void   fillMatrixConditional(int**matrix,  int rows,  int cols);
 void   fillMatrixIdentity(int**matrix,int rows,int cols);
 int    checkIdentity(int** matrix, int rows, int cols);
 int    traceMatrix(int** matrix,int rows, int cols);
